pythonProject/requirements.txt
aniso8601==8.1.0
bcrypt==3.2.0
between==0.2.7
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
et-xmlfile==1.0.1
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-WTF==0.14.3
future==0.18.2
idna==2.10
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.5
openpyxl==3.0.5
pandas==1.2.0
pwnedpasswords==2.0.0
pycparser==2.20
pyhibp==4.1.0
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.0
six==1.15.0
urllib3==1.26.2
websocket-client==0.57.0
Werkzeug==1.0.1
WTForms==2.3.3
pyjwt==2.0.1